requirements.txt
streamlit
scikit-learn
opencv-python
numpy
pillow
scikit-image
joblib
matplotlib
